requirements.txt
Django<4.0
django-compressor
django-crispy-forms<2.0
django-filter
django-guardian
django-sendfile
djangorestframework
jsonfield
Markdown
paramiko>=2.1.6
rcssmin
rjsmin
scandir
zipstream
django-cors-headers
psycopg2-binary
